Fix for discrete actions (#4181)

+++ ml-agents/mlagents/trainers/distributions_torch.py
@@ -42,7 +42,7 @@ class CategoricalDistInstance(nn.Module):
         return torch.multinomial(self.probs, 1)
 
     def pdf(self, value):
-        return torch.diag(self.probs.T[value.flatten()])
+        return torch.diag(self.probs.T[value.flatten().long()])
 
     def log_prob(self, value):
         return torch.log(self.pdf(value))
